+++ src/main/java/org/betonquest/betonquest/feature/registry/processor/CancelerProcessor.java
@@ -14,13 +14,13 @@ import org.betonquest.betonquest.id.ObjectiveID;
 import org.betonquest.betonquest.id.QuestCancelerID;
 import org.betonquest.betonquest.instruction.argument.IDArgument;
 import org.betonquest.betonquest.instruction.variable.location.VariableLocation;
+import org.betonquest.betonquest.message.ParsedSectionMessage;
 import org.betonquest.betonquest.quest.registry.processor.VariableProcessor;
 import org.betonquest.betonquest.variables.GlobalVariableResolver;
 import org.bukkit.configuration.ConfigurationSection;
 import org.jetbrains.annotations.Nullable;
 
 import java.lang.reflect.Array;
-import java.util.Map;
 
 /**
  * Stores Quest Canceler.
@@ -60,7 +60,7 @@ public class CancelerProcessor extends SectionProcessor<QuestCancelerID, QuestCa
 
     @Override
     protected QuestCanceler loadSection(final QuestPackage pack, final ConfigurationSection section) throws QuestException {
-        final Map<String, String> names = parseWithLanguage(pack, section, "name");
+        final ParsedSectionMessage names = new ParsedSectionMessage(variableProcessor, pack, section, "name");
         final String itemString = section.getString("item");
         final String rawItem = itemString == null ? pack.getConfig().getString("items.cancel_button") : itemString;
         final ItemID item = rawItem == null ? null : new ItemID(pack, rawItem);