+++ src/main/java/org/betonquest/betonquest/feature/registry/processor/CompassProcessor.java
@@ -7,11 +7,10 @@ import org.betonquest.betonquest.feature.QuestCompass;
 import org.betonquest.betonquest.id.CompassID;
 import org.betonquest.betonquest.id.ItemID;
 import org.betonquest.betonquest.instruction.variable.location.VariableLocation;
+import org.betonquest.betonquest.message.ParsedSectionMessage;
 import org.betonquest.betonquest.quest.registry.processor.VariableProcessor;
 import org.betonquest.betonquest.variables.GlobalVariableResolver;
 import org.bukkit.configuration.ConfigurationSection;
-
-import java.util.Map;
 
 /**
  * Loads and stores {@link QuestCompass}es.
@@ -36,7 +35,7 @@ public class CompassProcessor extends SectionProcessor<CompassID, QuestCompass> 
 
     @Override
     protected QuestCompass loadSection(final QuestPackage pack, final ConfigurationSection section) throws QuestException {
-        final Map<String, String> names = parseWithLanguage(pack, section, "name");
+        final ParsedSectionMessage names = new ParsedSectionMessage(variableProcessor, pack, section, "name");
         final String location = section.getString("location");
         if (location == null) {
             throw new QuestException("Location not defined");