+++ CHANGELOG.md
@@ -9,6 +9,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `quest compass` location now allows variables
 - `point` and `globalpoint` events and conditions now support variables
 - `cancel` event now has `bypass` keyword to ignore conditions in Quest Canceler
+- `quester` in conversation now supports variables
 ### Changed
 - message.yml file was deleted and instead the lang folder now contains all translations
 - menuConfig.yml file was deleted, translations are in the lang folder and the config is part of config.yml